+++ src/main/java/me/hsgamer/extrastorage/commands/subs/admin/AddCmd.java
@@ -35,9 +35,9 @@ public final class AddCmd
     @Override
     public void execute(CommandContext context) {
         String args0 = context.getArgs(0), args1 = context.getArgs(1);
-        int amount;
+        long amount;
         try {
-            amount = Digital.getBetween(1, Integer.MAX_VALUE, Integer.parseInt(args1));
+            amount = Digital.getBetween(1, Long.MAX_VALUE, Long.parseLong(args1));
         } catch (NumberFormatException ignored) {
             context.sendMessage(Message.getMessage("FAIL.not-number").replaceAll(VALUE_REGEX, args1));
             return;
@@ -63,7 +63,7 @@ public final class AddCmd
                     context.sendMessage(Message.getMessage("FAIL.storage-is-full"));
                     return;
                 }
-                if (amount > freeSpace) amount = (int) freeSpace;
+                if (amount > freeSpace) amount = freeSpace;
             }
             storage.add(args0, amount);
 
@@ -94,7 +94,7 @@ public final class AddCmd
                 context.sendMessage(Message.getMessage("FAIL.storage-is-full"));
                 return;
             }
-            if (amount > freeSpace) amount = (int) freeSpace;
+            if (amount > freeSpace) amount = freeSpace;
         }
         storage.add(args0, amount);
